requirements.txt

Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg
pandas
celery[redis]
redis
psycopg2-binary
openpyxl
flake8
black
isort
pre-commit
